// newcode/libs/modules/documents/shared/src/lib/model/entities/change-rubric-category-param-generated.interface.ts
/*
 * $Id$
 * Copyright(c) RIB Software GmbH
 */


export interface IChangeRubricCategoryParamGenerated {

/*
 * DocumentCategoryFk
 */
  DocumentCategoryFk?: number | null;

/*
 * DocumentId
 */
  DocumentId?: number | null;

/*
 * PrjDocumentTypeFk
 */
  PrjDocumentTypeFk?: number | null;

/*
 * RubricCategoryFk
 */
  RubricCategoryFk?: number | null;
}